initial fix: created packages and realocated the code
