+++ app/src/main/java/com/example/routinely/ui/components/GoogleLoginButton.kt
@@ -1,0 +1,38 @@
+package com.example.routinely.ui.components
+
+import androidx.compose.foundation.Image
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.height
+import androidx.compose.foundation.layout.width
+import androidx.compose.material3.Button
+import androidx.compose.material3.ButtonDefaults
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.res.painterResource
+import androidx.compose.ui.unit.dp
+import com.example.routinely.R
+
+@Composable
+fun GoogleLoginButton() {
+    Button(
+        onClick = {
+            // Ação executada quando o botão é clicado
+        },
+        modifier = Modifier
+            .fillMaxWidth(),
+        colors = ButtonDefaults.buttonColors(Color.White)
+    ) {
+        Image(
+            painter = painterResource(id = R.drawable.vector),
+            contentDescription = "Vector",
+            modifier = Modifier
+                .width(width = 18.dp)
+                .height(height = 18.dp)
+        )
+        Text(
+            text = "Entrar com Google", color = Color.Gray
+        )
+    }
+}
