+++ app/src/main/java/com/example/routinely/ui/components/LoginTextField.kt
@@ -1,0 +1,39 @@
+package com.example.routinely.ui.components
+
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.material3.ExperimentalMaterial3Api
+import androidx.compose.material3.OutlinedTextField
+import androidx.compose.material3.Text
+import androidx.compose.material3.TextFieldDefaults
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.getValue
+import androidx.compose.runtime.mutableStateOf
+import androidx.compose.runtime.saveable.rememberSaveable
+import androidx.compose.runtime.setValue
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.text.TextStyle
+
+@OptIn(ExperimentalMaterial3Api::class)
+@Composable
+fun LoginTextField(onEmailChange: (String) -> Unit) {
+    var text by rememberSaveable { mutableStateOf("") }
+    OutlinedTextField(
+        value = text,
+        onValueChange = {
+            text = it
+            onEmailChange(it)},
+        label = {
+            Text(
+                text = "Login",
+                style = TextStyle(color = Color.White) // Definindo a cor do texto como branco
+            )
+        },
+        colors = TextFieldDefaults.outlinedTextFieldColors(
+            focusedBorderColor = Color.White,
+            unfocusedBorderColor = Color.White
+        ),
+        modifier = Modifier
+            .fillMaxWidth() // Preencher toda a largura disponível no Row
+    )
+}
