+++ app/src/main/java/com/example/routinely/ui/components/PasswordTextField.kt
@@ -1,0 +1,43 @@
+package com.example.routinely.ui.components
+
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.text.KeyboardOptions
+import androidx.compose.material3.ExperimentalMaterial3Api
+import androidx.compose.material3.OutlinedTextField
+import androidx.compose.material3.Text
+import androidx.compose.material3.TextFieldDefaults
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.getValue
+import androidx.compose.runtime.mutableStateOf
+import androidx.compose.runtime.saveable.rememberSaveable
+import androidx.compose.runtime.setValue
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.text.TextStyle
+import androidx.compose.ui.text.input.KeyboardType
+import androidx.compose.ui.text.input.PasswordVisualTransformation
+
+@OptIn(ExperimentalMaterial3Api::class)
+@Composable
+fun PasswordTextField(onPasswordChange: (String) -> Unit) {
+    var password by rememberSaveable { mutableStateOf("") }
+    OutlinedTextField(
+        value = password,
+        onValueChange = {
+            password = it
+            onPasswordChange(it)},
+        label = {
+            Text(
+                text = "Senha",
+                style = TextStyle(color = Color.White) // Definindo a cor do texto como branco
+            )
+        },
+        colors = TextFieldDefaults.outlinedTextFieldColors(
+            focusedBorderColor = Color.White,
+            unfocusedBorderColor = Color.White
+        ),
+        visualTransformation = PasswordVisualTransformation(),
+        keyboardOptions = KeyboardOptions(keyboardType = KeyboardType.Password),
+        modifier = Modifier.fillMaxWidth() // Preencher toda a largura disponível no Row
+    )
+}
